requirements.txt
africastalking==1.1.4
AfricastalkingGateway==1.9
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.5
idna==2.7
python-decouple==3.1
pytz==2018.4
requests==2.19.0
schema==0.6.7
urllib3==1.23
django-appconf==1.0.2
django-bootstrap3==10.0.1
django-bootstrap4==0.0.6
django-imagekit==4.0.2
django-model-utils==3.1.2
gunicorn==19.8.1
pilkit==2.0
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
python-decouple==3.1
pytz==2018.4
six==1.11.0
whitenoise==3.3.1
astroid==1.6.4
Babel==2.5.1
coreapi==2.3.3
coreschema==0.0.4
django-braces==1.13.0
django-cors-headers==2.2.0
django-formtools==2.1
django-geojson==2.10.0
django-leaflet==0.22.0
django-map-widgets==0.1.9
django-phonenumber-field==1.3.0
django-private-chat==0.1.8
django-rest-swagger==2.1.2
django-simple-search==1.0.2
django-widget-tweaks==1.4.2
djangorestframework==3.6.4
djangorestframework-gis==0.11.2
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
openapi-codec==1.3.2
uritemplate==3.0.0
websockets==3.2
wrapt==1.10.11